feat: Add check for draging task into its child
This makes sure that the task being dragged is not being dropped into
one of it's own children

-Added hasChild() to treeNode which checks to see if the node argument
is a child
-Added the child check in TaskCommands.moveTask()

+++ src/taskCommands.js
@@ -123,6 +123,12 @@ export class TaskCommands {
     moveTask(movedTaskID, taskTargetID, position) {
         let movingTaskNode = TreeService.findNodeByID(this.root, movedTaskID);
 
+        let targetTaskNode = TreeService.findNodeByID(this.root, taskTargetID);
+
+        if(movingTaskNode.hasChild(targetTaskNode)) {
+            return undefined;
+        }
+
         TreeService.removeNodeByID(this.root, movedTaskID);
 
         if(position === "below") {
